Symfony Doctrine - script renamed (draft)
