+++ 04 Doctrine CRUD/composer.json
@@ -66,5 +66,7 @@
     },
     "require-dev": {
         "symfony/maker-bundle": "^1.29"
-    }
+    },
+    "name": "",
+    "description": ""
 }
