--- 04 Doctrine CRUD/src/Repository/GuestbookRepository.php
@@ -1,48 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\Guestbook;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method Guestbook|null find($id, $lockMode = null, $lockVersion = null)
- * @method Guestbook|null findOneBy(array $criteria, array $orderBy = null)
- * @method Guestbook[]    findAll()
- * @method Guestbook[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class GuestbookRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, Guestbook::class);
-    }
-
-    // /**
-    //  * @return Guestbook[] Returns an array of Guestbook objects
-    //  */
-    public function findByMsgField($value)
-    {
-        return $this->createQueryBuilder('g')
-            ->andWhere('g.msg LIKE :val')
-            ->setParameter('val', '%'.$value.'%')
-            ->orderBy('g.id', 'ASC')
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-
-
-    /*
-    public function findOneBySomeField($value): ?Guestbook
-    {
-        return $this->createQueryBuilder('g')
-            ->andWhere('g.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
-    }
-    */
-}
