+++ 04 Doctrine CRUD/src/Repository/NotesRepository.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Repository;
+
+use App\Entity\Notes;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @method Notes|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Notes|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Notes[]    findAll()
+ * @method Notes[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class NotesRepository extends ServiceEntityRepository
+{
+
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Notes::class);
+    }
+
+    // /**
+    //  * @return Notes[] Returns an array of Notes objects
+    //  */
+    public function findByMsgField($value)
+    {
+        return $this->createQueryBuilder('g')
+          ->andWhere('g.msg LIKE :val')
+          ->setParameter('val', '%' . $value . '%')
+          ->orderBy('g.id', 'ASC')
+          ->getQuery()
+          ->getResult();
+    }
+
+}
