+++ 04 Doctrine CRUD/templates/base.html.twig
@@ -6,7 +6,7 @@
         <meta name="description" content="">
         <meta name="author" content="">
 
-        <title>{% block title %}{% endblock %} | GuestBook</title>
+        <title>{% block title %}{% endblock %} | Notes</title>
 
         <link rel="stylesheet"
               href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
