+++ 04 Doctrine CRUD/templates/index.html.twig
@@ -1,6 +1,6 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Home{% endblock title %}
+{% block title %}Main{% endblock title %}
 
 {% block body %}
     <div class="bg-light text-dark border-10 border-info rounded p-3">
@@ -9,8 +9,8 @@
                 <div class="row justify-content-center py-2">
                     <form class="form col-10 col-sm-8 col-md-6 col-lg-4 align-self-center" action="/add" method="POST">
                         <div class="form__group form-group">
-                            <input name="nick" type="text" placeholder="nickname" class="form-control mb-1" autocomplete="off">
-                            <textarea name="msg" cols="30" rows="2" placeholder="message" class="form-control mt-1"></textarea>
+                            <input name="title" type="text" placeholder="Title" class="form-control mb-1" autocomplete="off">
+                            <textarea name="msg" cols="30" rows="2" placeholder="Message" class="form-control mt-1"></textarea>
                         </div>
                         <button type="submit" class="btn btn-info form-control">Submit</button>
                     </form>
