+++ 04 Doctrine CRUD/templates/records.html.twig
@@ -1,10 +1,11 @@
-Order <a href="?order={{ order == 'ASC' ? 'DESC' : 'ASC' }}">ASC/DESC</a>
+{% set sort = '?order=' ~ (order == 'ASC' ? 'DESC' : 'ASC') %}
+{% set order = '?order=' ~ order %}
 
-{{order}}
+Sort by Title: <a href="{{ sort }}">ASC/DESC</a>
 <hr>
 {% for record in records %}
     <p>
-        <b class="text-primary">{{ record.nick }}</b>
+        <b class="text-primary">{{ record.title }}</b>
         <span class="text-secondary">(#{{ record.id }} | {{ record.timestamp|date('Y.m.d H:i:s') }})</span>
         <br>
         <b>{{ record.msg }}</b>
@@ -20,7 +21,7 @@ Order <a href="?order={{ order == 'ASC' ? 'DESC' : 'ASC' }}">ASC/DESC</a>
 
         {% if previous_page %}
             <li class="page-item">
-                <a class="page-link" href="/{{ previous_page }}">Previous</a>
+                <a class="page-link" href="/{{ previous_page }}{{ order }}">Previous</a>
             </li>
         {% else %}
             <li class="page-item disabled">
@@ -37,7 +38,7 @@ Order <a href="?order={{ order == 'ASC' ? 'DESC' : 'ASC' }}">ASC/DESC</a>
 
         {% if next_page %}
             <li class="page-item">
-                <a class="page-link" href="/{{ next_page }}">Next</a>
+                <a class="page-link" href="/{{ next_page }}{{ order }}">Next</a>
             </li>
         {% else %}
             <li class="page-item disabled">
